fix the pressure test script
