+++ src/test/java/org/fisco/bcos/channel/test/contract/PerfomanceOk.java
@@ -16,7 +16,6 @@ import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import java.util.Random;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
@@ -27,7 +26,7 @@ import static java.lang.System.exit;
 public class PerfomanceOk {
 	static Logger logger = LoggerFactory.getLogger(PerfomanceOk.class);
 	private static AtomicInteger sended = new AtomicInteger(0);
-	
+
 	public static void main(String[] args) throws Exception {
 		String groupId = args[3];
 		//初始化Service
@@ -37,12 +36,11 @@ public class PerfomanceOk {
 
 		System.out.println("开始测试...");
 		System.out.println("===================================================================");
-		
+
 		ChannelEthereumService channelEthereumService = new ChannelEthereumService();
 		channelEthereumService.setChannelService(service);
 
-		ScheduledExecutorService scheduledExecutorService =
-				Executors.newScheduledThreadPool(500);
+		ScheduledExecutorService scheduledExecutorService = Executors.newScheduledThreadPool(500);
 		Web3j web3 = Web3j.build(channelEthereumService,  15 * 100, scheduledExecutorService);
 
 		//初始化交易签名私钥
@@ -53,20 +51,20 @@ public class PerfomanceOk {
 		BigInteger gasPrice = new BigInteger("30000000");
 		BigInteger gasLimit = new BigInteger("30000000");
 		BigInteger initialWeiValue = new BigInteger("0");
-		
+
 		//解析参数
 		String command = args[0];
 		Integer count = 0;
 		Integer qps = 0;
 		Integer startNum = 0;
-		
+
 		ThreadPoolTaskExecutor threadPool = new ThreadPoolTaskExecutor();
 		threadPool.setCorePoolSize(2000);
 		threadPool.setMaxPoolSize(2000);
-		threadPool.setQueueCapacity(50000);
-		
+		threadPool.setQueueCapacity(100000);
+
 		threadPool.initialize();
-		
+
 		System.out.println("部署合约");
 		Ok ok = Ok.deploy(web3, credentials, gasPrice, gasLimit, initialWeiValue).send();
 
@@ -81,7 +79,7 @@ public class PerfomanceOk {
 
 		PerfomanceOkCallback callback = new PerfomanceOkCallback();
 		callback.setTotal(count);
-		
+
 		RateLimiter limiter = RateLimiter.create(qps);
 		Integer area = count / 10;
 
@@ -101,49 +99,30 @@ public class PerfomanceOk {
 
 					Long currentTime = System.currentTimeMillis();
 
-							String userName = String.valueOf("User " + String.valueOf(seq + start));
-							try {
-								CompletableFuture<TransactionReceipt> future = ok.trans(new BigInteger("4")).sendAsync();
-								threadArray.add(future);
-							} catch (Exception e) {
-								logger.info(e.getMessage());
-							}
-
-							int current = sended.incrementAndGet();
-
-							if (current >= area && ((current % area) == 0)) {
-								System.out.println("已发送: " + current + "/" + total + " 交易");
-								//	System.out.println("耗时 ms" + Long.toString(System.currentTimeMillis() - currentTime));
-							}
+					String userName = String.valueOf("User " + String.valueOf(seq + start));
+					try {
+						CompletableFuture<TransactionReceipt> future = ok.trans(new BigInteger("4")).sendAsync();
+						TransactionReceipt transactionReceipt = future.get();
+						callback.onResponse(System.currentTimeMillis() - currentTime);
+					} catch (Exception e) {
+						logger.info(e.getMessage());
 					}
+
+					int current = sended.incrementAndGet();
+
+					if (current >= area && ((current % area) == 0)) {
+						System.out.println("已发送: " + current + "/" + total + " 交易");
+						//	System.out.println("耗时 ms" + Long.toString(System.currentTimeMillis() - currentTime));
+					}
+				}
 
 
 			});
 		}
-	int count1 = 1;
 
-		while(threadArray.size()>0) {
-			for (int i = 0; i < threadArray.size(); i++) {
-
-				if (threadArray.get(i).isDone()) {
-					threadArray.remove(threadArray.get(i));
-					count1++;
-				}
-
-			if(count1 %2000 ==0) {
-				Long time = System.currentTimeMillis() - currentTime;
-				System.out.println("2000笔交易耗时 ms" + Long.toString(System.currentTimeMillis() - currentTime));
-				double tps = 2000.0/(time/1000.0);
-				System.out.println("tps 是" + (double)Math.round(tps*100)/100);
-				currentTime= System.currentTimeMillis();
-			}
-
-			}
-			Thread.sleep(500);
-		}
-
-		System.out.println("已收到交易 " + (count1-1));
-		exit(1);
+		Thread.sleep(5000);
+		System.out.println("全部交易已发送: " + count);
+		exit(0);
 
 	}
 }
